requirements.txt
Django~=2.2.10
python-dotenv~=0.14.0
django-crispy-forms~=1.9.2
djangorestframework~=3.12.1